add new class HeroHealthBar and HeroXPBar and fix NullPointerException
* add totalXP to XPComponent

* xpPopups are now calculated with the new totalXP

* move methods for the hero's healthbar in HeroHealthBar class

* move methods for the hero's xpBar in HeroXPBar class

* rename xpBar assets to match healthBar assets

* health- and xpbar is now only created if the associated component is present
